requirements.txt
# Deep Learning y Procesamiento de Imágenes
ultralytics==8.0.196
torch>=2.0.0
torchvision>=0.15.0
numpy==1.24.3
opencv-python==4.8.1.78
Pillow==10.0.0
matplotlib==3.7.2

# GUI
tk==0.1.0

# Testing
pytest==7.4.2

scikit_image ==0.25.2